fix license lines

+++ ffmulticonverter/progress.py
@@ -4,8 +4,8 @@
 # Copyright (C) 2011-2012 Ilias Stamatis <stamatis.iliass@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
-# the Free Software Foundation, either version 3 of the License, or
 # it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
 #
 # This program is distributed in the hope that it will be useful,
@@ -59,7 +59,7 @@ class Progress(QDialog):
 
         self.files = files
         self.delete = delete
-        self.step = 100 / len(files)
+        self.step = int(100 / len(files))
         self.ok = 0
         self.error = 0
         self.running = True
